Move all relay files under common folder
